fix(auth): relocate stranded verify*Auth helpers + repair inline require paths (#164)
Follow-up to 5c24855. The original splitter only tracked top-level
`exports.NAME = ...` declarations, so five non-exported helpers between
`registerAuth` and `verifyAuth` in the source file (verifyGithubAccessToken,
verifyGithubAuth, verifyGoogleIdToken, verifyGoogleAuth, verifyLocalAuth)
got swallowed into `registerAuth`'s line range and ended up in
`controller/register.js`, while their only caller — `verifyAuth` — landed
in `controller/authHelpers.js`. Every login attempt threw
`ReferenceError: verifyLocalAuth is not defined` at runtime.

The previous smoke test missed it because it only checked
`typeof exports.verifyAuth === 'function'`, which is true even when calling
the function blows up.

Changes:
- Move the 5 verify*Auth helpers into authHelpers.js, placed before
  verifyAuth so their definitions are in scope at call time.
- Patch inline `require()` paths that the preamble-rewriter didn't touch
  (they live inside function bodies, not at file top):
  * authHelpers.js: `../Template/passwordExpiredMail` -> `../../Template/passwordExpiredMail`
  * authHelpers.js: `../Template/forgotPassword`     -> `../../Template/forgotPassword`
  * Drop the inline `const logger = require("../../Config/loggerConfig")`
    inside verifyGoogleAuth — it was already redundant in the original
    (logger is imported at the top) and now also pointed to a non-existent
    depth.
- register.js trimmed to 31 lines (preamble + registerAuth only).

Verification:
- New smoke test actually invokes `verifyAuth` with each authProvider
  ('local', 'github', 'google') and asserts the early-return validation
  message — proving every verify* helper symbol resolves at runtime.
- Verified the two inline template require specifiers resolve to real
  files via `require.resolve`.
- Grep-confirmed no stale `../Template/` or `../../Config/` paths
  survive in any Auth sub-file.
- `npm test` 20/20 passing.
